Move HIRMS utility functions to a trait, make them available in HIRMG
