+++ lib/Fhp/Segment/HIRMS/HIRMSv2.php
@@ -22,21 +22,8 @@ use Fhp\Segment\BaseSegment;
  */
 class HIRMSv2 extends BaseSegment
 {
+    use RueckmeldungContainer; // For utility functions.
+
     /** @var Rueckmeldung[] @Max(99) */
     public $rueckmeldung;
-
-    /**
-     * @param int $code The value of Rueckmeldung.rueckmeldungscode to search for.
-     * @return Rueckmeldung|null The corresponding Rueckmeldung instance, or null if not found.
-     */
-    public function findRueckmeldung($code)
-    {
-        $matches = array_values(array_filter($this->rueckmeldung, function ($rueckmeldung) use ($code) {
-            return $rueckmeldung->rueckmeldungscode === $code;
-        }));
-        if (count($matches) > 1) {
-            throw new \InvalidArgumentException("Unexpectedly multiple matches for Rueckmeldungscode $code");
-        }
-        return empty($matches) ? null : $matches[0];
-    }
 }
